requirements.txt
librosa
numpy
scikit-learn
kafka-python
confluent-kafka